cache val with hash as key, and fix OP_POLY_COPY_FROM_MEM's cache query

+++ include/brepvm/CodeBlock.h
@@ -1,11 +1,13 @@
 #pragma once
 
+#include <stdint.h>
+
 namespace brepvm
 {
 
 struct CodeBlock
 {
-	size_t hash;
+	uint32_t hash;
 
 	int begin;
 	int end;
